src/competitor_layer/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "competitor-layer"
version = "0.1.0"
description = "Agnes Competitor Layer - supplier discovery for CPG ingredient sourcing"
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2.0,<3.0",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "google-genai>=1.0",
]

[tool.setuptools.packages.find]
include = ["competitor_layer*"]

[project.scripts]
competitor-layer = "competitor_layer.cli:main"
